Refactor: move mockAccount helpers to own file
